+++ src/heads/molecular/__init__.py
@@ -1,0 +1,1 @@
+"""Molecular head package. The head instance is exposed as ``head.HEAD``."""